refactor: #149 - 즐겨찾기 스페이스 검색 fetch 함수 파일명 수정
